+++ src/org/usfirst/frc/team246/robot/OverclockedLibraries/AnalogIn.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team246.robot;
+package org.usfirst.frc.team246.robot.OverclockedLibraries;
 
 import edu.wpi.first.wpilibj.AnalogInput;
 import edu.wpi.first.wpilibj.PIDSource;
