+++ src/org/usfirst/frc/team246/robot/OverclockedLibraries/DataInterpolator.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team246.robot;
+package org.usfirst.frc.team246.robot.OverclockedLibraries;
 
 /**
  * This is a generic class for the interpolation of data stored in a 2D array.
