+++ src/org/usfirst/frc/team246/robot/OverclockedLibraries/Diagnostics.java
@@ -1,6 +1,8 @@
-package org.usfirst.frc.team246.robot;
+package org.usfirst.frc.team246.robot.OverclockedLibraries;
 
 import java.util.ArrayList;
+
+import org.usfirst.frc.team246.robot.CANTalonPotentiometer;
 
 import edu.wpi.first.wpilibj.CANTalon;
 import edu.wpi.first.wpilibj.Encoder;
