+++ src/org/usfirst/frc/team246/robot/OverclockedLibraries/VectorPIDSource.java
@@ -1,0 +1,7 @@
+package org.usfirst.frc.team246.robot.OverclockedLibraries;
+
+import org.usfirst.frc.team246.robot.Vector2D;
+
+public interface VectorPIDSource {
+	public Vector2D pidGet();
+}
