--- src/org/usfirst/frc/team246/robot/VectorPIDOutput.java
@@ -1,5 +1,0 @@
-package org.usfirst.frc.team246.robot;
-
-public interface VectorPIDOutput {
-	public void pidWrite(Vector2D vector);
-}
