--- src/org/usfirst/frc/team246/robot/VectorPIDSource.java
@@ -1,5 +1,0 @@
-package org.usfirst.frc.team246.robot;
-
-public interface VectorPIDSource {
-	public Vector2D pidGet();
-}
