+++ src/org/usfirst/frc/team246/robot/commands/GoFast.java
@@ -1,21 +1,23 @@
-package org.usfirst.frc.team246.robot;
+package org.usfirst.frc.team246.robot.commands;
+
+import org.usfirst.frc.team246.robot.Robot;
+import org.usfirst.frc.team246.robot.OverclockedLibraries.AlertMessage;
+import org.usfirst.frc.team246.robot.OverclockedLibraries.UdpAlertService;
 
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
  *
  */
-public class ZeroNavX extends Command {
+public class GoFast extends Command {
 
-	double offset;
-	
-    public ZeroNavX(double offset) {
-    	this.offset = offset;
+    public GoFast() {
     }
 
     // Called just before this Command runs the first time
     protected void initialize() {
-    	Robot.drivetrain.navX.zeroYaw(offset);
+    	Robot.drivetrain.setMaxSpeed(Robot.FAST_CRAB_SPEED, Robot.FAST_SPIN_SPEED);
+    	UdpAlertService.sendAlert(new AlertMessage("Going fast"));
     }
 
     // Called repeatedly when this Command is scheduled to run
@@ -24,15 +26,18 @@ public class ZeroNavX extends Command {
 
     // Make this return true when this Command no longer needs to run execute()
     protected boolean isFinished() {
-        return true;
+        return false;
     }
 
     // Called once after isFinished returns true
     protected void end() {
+    	Robot.drivetrain.setMaxSpeed(Robot.SLOW_CRAB_SPEED, Robot.SLOW_SPIN_SPEED);
+    	UdpAlertService.sendAlert(new AlertMessage("Slowing Down"));
     }
 
     // Called when another command which requires one or more of the same
     // subsystems is scheduled to run
     protected void interrupted() {
+    	end();
     }
 }
