+++ src/org/usfirst/frc/team246/robot/commands/ZeroNavX.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team246.robot;
+package org.usfirst.frc.team246.robot.commands;
 
 import org.usfirst.frc.team246.robot.Robot;
 
@@ -7,15 +7,17 @@ import edu.wpi.first.wpilibj.command.Command;
 /**
  *
  */
-public class GoFast extends Command {
+public class ZeroNavX extends Command {
 
-    public GoFast() {
+	double offset;
+	
+    public ZeroNavX(double offset) {
+    	this.offset = offset;
     }
 
     // Called just before this Command runs the first time
     protected void initialize() {
-    	Robot.drivetrain.setMaxSpeed(Robot.FAST_CRAB_SPEED, Robot.FAST_SPIN_SPEED);
-    	UdpAlertService.sendAlert(new AlertMessage("Going fast"));
+    	Robot.drivetrain.navX.zeroYaw(offset);
     }
 
     // Called repeatedly when this Command is scheduled to run
@@ -24,18 +26,15 @@ public class GoFast extends Command {
 
     // Make this return true when this Command no longer needs to run execute()
     protected boolean isFinished() {
-        return false;
+        return true;
     }
 
     // Called once after isFinished returns true
     protected void end() {
-    	Robot.drivetrain.setMaxSpeed(Robot.SLOW_CRAB_SPEED, Robot.SLOW_SPIN_SPEED);
-    	UdpAlertService.sendAlert(new AlertMessage("Slowing Down"));
     }
 
     // Called when another command which requires one or more of the same
     // subsystems is scheduled to run
     protected void interrupted() {
-    	end();
     }
 }
